test(factory): place GP-023 isolation in owned contract
Move the real child-process environment assertion into the canonical factory-stack root-binding contract suite recognized by the task-platform ownership policy. The behavior is unchanged: live database, auth, OpenClaw, feature flags, and trusted-delivery settings remain unobservable in GP-023 children.

Verify the exact PR diff against the change-completeness gate before publishing.

Refs #373
